API

Types and constants

Enzyme.Active

Enzyme.Annotation

Enzyme.Const

Enzyme.Duplicated

Functions and macros

Enzyme.autodiff

Enzyme.autodiff

Enzyme.autodiff_deferred

Enzyme.autodiff_deferred

Documentation

Enzyme.Active — Type

struct Active{T} <: Annotation{T}

Constructor: Active(x)

Mark a function argument x of autodiff as active, Enzyme will auto-differentiate in respect Active arguments.

Note

Enzyme gradients with respect to integer values are zero. Active will automatically convert plain integers to floating point values, but cannot do so for integer values in tuples and structs.

Enzyme.Annotation — Type

abstract type Annotation{T}

Abstract type for autodiff function argument wrappers like Const, Active and Duplicated.

Enzyme.Const — Type

struct Const{T} <: Annotation{T}

Constructor: Const(x)

Mark a function argument x of autodiff as constant, Enzyme will not auto-differentiate in respect Const arguments.

Enzyme.Duplicated — Type

struct Duplicated{T} <: Annotation{T}

Constructor: Duplicated(x, ∂f_∂x)

Mark a function argument x of autodiff as duplicated, Enzyme will auto-differentiate in respect to such arguments, with dx acting as an accumulator for gradients (so $\partial f / \partial x$ will be added to) ∂f_∂x.

Enzyme.autodiff — Method

autodiff(f, Activity, args...)

Auto-differentiate function f at arguments args.

Limitations:

f may only return a Real (of a built-in/primitive type) or nothing, not an array, struct, BigFloat, etc. To handle vector-valued return types, use a mutating f! that returns nothing and stores it's return value in one of the arguments, which must be wrapped in a Duplicated.

f may not allocate memory, this restriction is likely to be removed in future versions. Technically it can currently allocate memory directly, but not in a function called by f.

args may be numbers, arrays, structs of numbers, structs of arrays and so on. Enzyme will only differentiate in respect to arguments that are wrapped in an Active (for immutable arguments like primitive types and structs thereof) or Duplicated (for mutable arguments like arrays, Refs and structs thereof). Non-annotated arguments will automatically be treated as Const.

Activity is the Activity of the return value, it may be Const or Active.

Example:

using Enzyme

a = 4.2
b = [2.2, 3.3]; ∂f_∂b = zero(b)
c = 55; d = 9

f(a, b, c, d) = a * √(b[1]^2 + b[2]^2) + c^2 * d^2
∂f_∂a, ∂f_∂d = autodiff(f, Active, Active(a), Duplicated(b, ∂f_∂b), c, Active(d))

# output

(3.966106403010388, 54450.0)

here, autodiff returns a tuple $(\partial f/\partial a, \partial f/\partial d)$, while $\partial f/\partial b$ will be added to ∂f_∂b (but not returned). c will be treated as Const(c).

Note

Enzyme gradients with respect to integer values are zero. Active will automatically convert plain integers to floating point values, but cannot do so for integer values in tuples and structs.

Enzyme.autodiff — Method

autodiff(f, args...)

Like autodiff but will try to guess the activity of the return value.

Enzyme.autodiff_deferred — Method

autodiff_deferred(f, Activity, args...)

Same as autodiff but uses deferred compilation to support usage in GPU code, as well as high-order differentiation.

Enzyme.autodiff_deferred — Method

autodiff_deferred(f, args...)

Like autodiff_deferred but will try to guess the activity of the return value.