API

Types and constants

EnzymeCore.Active

EnzymeCore.Annotation

EnzymeCore.BatchDuplicated

EnzymeCore.BatchDuplicatedNoNeed

EnzymeCore.Const

EnzymeCore.Duplicated

EnzymeCore.DuplicatedNoNeed

EnzymeCore.ForwardMode

EnzymeCore.Mode

EnzymeCore.ReverseMode

EnzymeCore.ReverseModeSplit

EnzymeCore.EnzymeRules.AugmentedReturn

EnzymeCore.EnzymeRules.Config

Functions and macros

Enzyme.gradient

Enzyme.gradient

Enzyme.gradient

Enzyme.gradient!

Enzyme.jacobian

Enzyme.jacobian

EnzymeCore.autodiff

EnzymeCore.autodiff

EnzymeCore.autodiff

EnzymeCore.autodiff

EnzymeCore.autodiff_deferred

EnzymeCore.autodiff_deferred

EnzymeCore.autodiff_deferred

EnzymeCore.autodiff_deferred_thunk

EnzymeCore.autodiff_thunk

EnzymeCore.autodiff_thunk

EnzymeCore.EnzymeRules.augmented_primal

EnzymeCore.EnzymeRules.forward

EnzymeCore.EnzymeRules.inactive

EnzymeCore.EnzymeRules.reverse

Documentation

Enzyme.gradient! — Method

gradient!(::ReverseMode, dx, f, x)

Compute the gradient of an array-input function f using reverse mode, storing the derivative result in an existing array dx.

Example:

f(x) = x[1]*x[2]

dx = [0.0, 0.0]
gradient!(Reverse, dx, f, [2.0, 3.0])

# output

2-element Vector{Float64}:
 3.0
 2.0

Enzyme.gradient — Method

gradient(::ForwardMode, f, x; shadow=onehot(x))

Compute the gradient of an array-input function f using forward mode. The optional keyword argument shadow is a vector of one-hot vectors of type x which are used to forward-propagate into the return. For performance reasons, this should be computed once, outside the call to gradient, rather than within this call.

Example:

f(x) = x[1]*x[2]

grad = gradient(Forward, f, [2.0, 3.0])

# output

(3.0, 2.0)

Enzyme.gradient — Method

gradient(::ReverseMode, f, x)

Compute the gradient of an array-input function f using reverse mode. This will allocate and return new array with the gradient result.

Example:

f(x) = x[1]*x[2]

grad = gradient(Reverse, f, [2.0, 3.0])

# output

2-element Vector{Float64}:
 3.0
 2.0

Enzyme.gradient — Method

gradient(::ForwardMode, f, x, ::Val{chunk}; shadow=onehot(x))

Compute the gradient of an array-input function f using vector forward mode. Like gradient, except it uses a chunk size of chunk to compute chunk derivatives in a single call.

Example:

f(x) = x[1]*x[2]

grad = gradient(Forward, f, [2.0, 3.0], Val(2))

# output

(3.0, 2.0)

Enzyme.jacobian — Method

jacobian(::ForwardMode, f, x; shadow=onehot(x))
jacobian(::ForwardMode, f, x, ::Val{chunk}; shadow=onehot(x))

Compute the jacobian of an array-input function f using (potentially vector) forward mode. This is a simple rename of the gradient function, and all relevant arguments apply here.

Example:

f(x) = [x[1]*x[2], x[2]]

grad = jacobian(Forward, f, [2.0, 3.0])

# output

2×2 Matrix{Float64}:
 3.0  2.0
 0.0  1.0

Enzyme.jacobian — Method

jacobian(::ReverseMode, f, x, ::Val{num_outs}, ::Val{chunk})

Compute the jacobian of an array-input function f using (potentially vector) reverse mode. The chunk argument denotes the chunk size to use and num_outs denotes the number of outputs f will return in an array.

Example:

f(x) = [x[1]*x[2], x[2]]

grad = jacobian(Reverse, f, [2.0, 3.0], Val(2))

# output

2×2 Matrix{Float64}:
 3.0  2.0
 0.0  1.0

EnzymeCore.autodiff — Method

autodiff(::ForwardMode, f, Activity, args...)

Auto-differentiate function f at arguments args using forward mode.

args may be numbers, arrays, structs of numbers, structs of arrays and so on. Enzyme will only differentiate in respect to arguments that are wrapped in a Duplicated or similar argument. Non-annotated arguments will automatically be treated as Const. Unlike reverse mode in autodiff, Active arguments are not allowed here, since all

Activity is the Activity of the return value, it may be:

Const if the return is not to be differentiated with respect to

Duplicated, if the return is being differentiated with respect to and both the original value and the derivative return are desired

DuplicatedNoNeed, if the return is being differentiated with respect to and only the derivative return is desired.

Example returning both original return and derivative:

a = 4.2
b = [2.2, 3.3]; ∂f_∂b = zero(b)
c = 55; d = 9

f(x) = x*x
res, ∂f_∂x = autodiff(Forward, f, Duplicated, Duplicated(3.14, 1.0))

# output

(9.8596, 6.28)

Example returning just the derivative:

a = 4.2
b = [2.2, 3.3]; ∂f_∂b = zero(b)
c = 55; d = 9

f(x) = x*x
∂f_∂x = autodiff(Forward, f, DuplicatedNoNeed, Duplicated(3.14, 1.0))

# output

(6.28,)

EnzymeCore.autodiff — Method

autodiff(mode::Mode, f, args...)

Like autodiff but will try to guess the activity of the return value.

EnzymeCore.autodiff — Method

autodiff(mode::Mode, f, ::Type{A}, args...)

Like autodiff but will try to extend f to an annotation, if needed.

EnzymeCore.autodiff — Method

autodiff(::ReverseMode, f, Activity, args...)

Auto-differentiate function f at arguments args using reverse mode.

Limitations:

f may only return a Real (of a built-in/primitive type) or nothing, not an array, struct, BigFloat, etc. To handle vector-valued return types, use a mutating f! that returns nothing and stores it's return value in one of the arguments, which must be wrapped in a Duplicated.

args may be numbers, arrays, structs of numbers, structs of arrays and so on. Enzyme will only differentiate in respect to arguments that are wrapped in an Active (for immutable arguments like primitive types and structs thereof) or Duplicated (for mutable arguments like arrays, Refs and structs thereof). Non-annotated arguments will automatically be treated as Const.

Activity is the Activity of the return value, it may be Const or Active.

Example:

a = 4.2
b = [2.2, 3.3]; ∂f_∂b = zero(b)
c = 55; d = 9

f(a, b, c, d) = a * √(b[1]^2 + b[2]^2) + c^2 * d^2
∂f_∂a, _, _, ∂f_∂d = autodiff(Reverse, f, Active, Active(a), Duplicated(b, ∂f_∂b), c, Active(d))[1]

# output

(3.966106403010388, nothing, nothing, 54450.0)

here, autodiff returns a tuple $(\partial f/\partial a, \partial f/\partial d)$, while $\partial f/\partial b$ will be added to ∂f_∂b (but not returned). c will be treated as Const(c).

One can also request the original returned value of the computation.

Example:

Enzyme.autodiff(ReverseWithPrimal, x->x*x, Active(3.0))

# output

((6.0,), 9.0)

Note

Enzyme gradients with respect to integer values are zero. Active will automatically convert plain integers to floating point values, but cannot do so for integer values in tuples and structs.

EnzymeCore.autodiff_deferred — Method

autodiff_deferred(::ForwardMode, f, Activity, args...)

Same as autodiff(::ForwardMode, ...) but uses deferred compilation to support usage in GPU code, as well as high-order differentiation.

EnzymeCore.autodiff_deferred — Method

autodiff_deferred(mode::Mode, f, ::Type{A}, args...)

Like autodiff_deferred but will try to extend f to an annotation, if needed.

EnzymeCore.autodiff_deferred — Method

autodiff_deferred(::ReverseMode, f, Activity, args...)

Same as autodiff but uses deferred compilation to support usage in GPU code, as well as high-order differentiation.

EnzymeCore.autodiff_deferred_thunk — Method

autodiff_deferred_thunk(::ReverseModeSplit, ftype, Activity, argtypes...)

Provide the split forward and reverse pass functions for annotated function type ftype when called with args of type argtypes when using reverse mode.

Activity is the Activity of the return value, it may be Const, Active, or Duplicated (or its variants DuplicatedNoNeed, BatchDuplicated, and BatchDuplicatedNoNeed).

The forward function will return a tape, the primal (or nothing if not requested), and the shadow (or nothing if not a Duplicated variant), and tapes the corresponding type arguements provided.

The reverse function will return the derivative of Active arguments, updating the Duplicated arguments in place. The same arguments to the forward pass should be provided, followed by the adjoint of the return (if the return is active), and finally the tape from the forward pass.

Example:

A = [2.2]; ∂A = zero(A)
v = 3.3

function f(A, v)
    res = A[1] * v
    A[1] = 0
    res
end

forward, reverse = autodiff_deferred_thunk(ReverseSplitWithPrimal, Const{typeof(f)}, Active, Duplicated{typeof(A)}, Active{typeof(v)})

tape, result, shadow_result  = forward(Const(f), Duplicated(A, ∂A), Active(v))
_, ∂v = reverse(Const(f), Duplicated(A, ∂A), Active(v), 1.0, tape)[1]

result, ∂v, ∂A 

# output

(7.26, 2.2, [3.3])

EnzymeCore.autodiff_thunk — Method

autodiff_thunk(::ForwardMode, ftype, Activity, argtypes...)

Provide the thunk forward mode function for annotated function type ftype when called with args of type argtypes.

Activity is the Activity of the return value, it may be Const or Duplicated (or its variants DuplicatedNoNeed, BatchDuplicated, andBatchDuplicatedNoNeed).

The forward function will return the primal (if requested) and the shadow (or nothing if not a Duplicated variant).

Example returning both original return and derivative:

a = 4.2
b = [2.2, 3.3]; ∂f_∂b = zero(b)
c = 55; d = 9

f(x) = x*x
forward = autodiff_thunk(Forward, Const{typeof(f)}, Duplicated, Duplicated{Float64})
res, ∂f_∂x = forward(Const(f), Duplicated(3.14, 1.0))

# output

(9.8596, 6.28)

Example returning just the derivative:

a = 4.2
b = [2.2, 3.3]; ∂f_∂b = zero(b)
c = 55; d = 9

f(x) = x*x
forward = autodiff_thunk(Forward, Const{typeof(f)}, DuplicatedNoNeed, Duplicated{Float64})
∂f_∂x = forward(Const(f), Duplicated(3.14, 1.0))

# output

(6.28,)

EnzymeCore.autodiff_thunk — Method

autodiff_thunk(::ReverseModeSplit, ftype, Activity, argtypes...)

Provide the split forward and reverse pass functions for annotated function type ftype when called with args of type argtypes when using reverse mode.

Activity is the Activity of the return value, it may be Const, Active, or Duplicated (or its variants DuplicatedNoNeed, BatchDuplicated, and BatchDuplicatedNoNeed).

The forward function will return a tape, the primal (or nothing if not requested), and the shadow (or nothing if not a Duplicated variant), and tapes the corresponding type arguements provided.

The reverse function will return the derivative of Active arguments, updating the Duplicated arguments in place. The same arguments to the forward pass should be provided, followed by the adjoint of the return (if the return is active), and finally the tape from the forward pass.

Example:

A = [2.2]; ∂A = zero(A)
v = 3.3

function f(A, v)
    res = A[1] * v
    A[1] = 0
    res
end

forward, reverse = autodiff_thunk(ReverseSplitWithPrimal, Const{typeof(f)}, Active, Duplicated{typeof(A)}, Active{typeof(v)})

tape, result, shadow_result  = forward(Const(f), Duplicated(A, ∂A), Active(v))
_, ∂v = reverse(Const(f), Duplicated(A, ∂A), Active(v), 1.0, tape)[1]

result, ∂v, ∂A 

# output

(7.26, 2.2, [3.3])

EnzymeCore.Active — Type

Active(x)

Mark a function argument x of autodiff as active, Enzyme will auto-differentiate in respect Active arguments.

Note

Enzyme gradients with respect to integer values are zero. Active will automatically convert plain integers to floating point values, but cannot do so for integer values in tuples and structs.

EnzymeCore.Annotation — Type

abstract type Annotation{T}

Abstract type for autodiff function argument wrappers like Const, Active and Duplicated.

EnzymeCore.BatchDuplicated — Type

BatchDuplicated(x, ∂f_∂xs)

Like Duplicated, except contains several shadows to compute derivatives for all at once. Argument ∂f_∂xs should be a tuple of the several values of type x.

EnzymeCore.BatchDuplicatedNoNeed — Type

BatchDuplicatedNoNeed(x, ∂f_∂xs)

Like DuplicatedNoNeed, except contains several shadows to compute derivatives for all at once. Argument ∂f_∂xs should be a tuple of the several values of type x.

EnzymeCore.Const — Type

Const(x)

Mark a function argument x of autodiff as constant, Enzyme will not auto-differentiate in respect Const arguments.

EnzymeCore.Duplicated — Type

Duplicated(x, ∂f_∂x)

Mark a function argument x of autodiff as duplicated, Enzyme will auto-differentiate in respect to such arguments, with dx acting as an accumulator for gradients (so $\partial f / \partial x$ will be added to) ∂f_∂x.

EnzymeCore.DuplicatedNoNeed — Type

DuplicatedNoNeed(x, ∂f_∂x)

Like Duplicated, except also specifies that Enzyme may avoid computing the original result and only compute the derivative values.

EnzymeCore.ForwardMode — Type

struct Forward <: Mode

Forward mode differentiation

EnzymeCore.Mode — Type

abstract type Mode

Abstract type for what differentiation mode will be used.

EnzymeCore.ReverseMode — Type

struct ReverseMode{ReturnPrimal} <: Mode

Reverse mode differentiation.

ReturnPrimal: Should Enzyme return the primal return value from the augmented-forward.

EnzymeCore.ReverseModeSplit — Type

struct ReverseModeSplit{ReturnPrimal,ReturnShadow,Width,ModifiedBetween} <: Mode

Reverse mode differentiation.

ReturnPrimal: Should Enzyme return the primal return value from the augmented-forward.

ReturnShadow: Should Enzyme return the shadow return value from the augmented-forward.

Width: Batch Size (0 if to be automatically derived)

ModifiedBetween: Tuple of each argument's modified between state (true if to be automatically derived).

EnzymeCore.EnzymeRules.AugmentedReturn — Type

AugmentedReturn(primal, shadow, tape)

Augment the primal return value of a function with its shadow, as well as any additional information needed to correctly compute the reverse pass, stored in tape.

Unless specified by the config that a variable is not overwritten, rules must assume any arrays/data structures/etc are overwritten between the forward and the reverse pass. Any floats or variables passed by value are always preserved as is (as are the arrays themselves, just not necessarily the values in the array).

See also augmented_primal.

EnzymeCore.EnzymeRules.Config — Type

Config{NeedsPrimal, NeedsShadow, Width, Overwritten}
ConfigWidth{Width} = Config{<:Any,<:Any, Width}

Configuration type to dispatch on in custom reverse rules (see augmented_primal and reverse).

NeedsPrimal and NeedsShadow: boolean values specifying whether the primal and shadow (resp.) should be returned.

Width: an integer that specifies the number of adjoints/shadows simultaneously being propagated.

Overwritten: a tuple of booleans of whether each argument (including the function itself) is modified between the forward and reverse pass (true if potentially modified between).

Getters for the four type parameters are provided by needs_primal, needs_shadow, width, and overwritten.

EnzymeCore.EnzymeRules.augmented_primal — Function

augmented_primal(::Config, func::Annotation{typeof(f)}, RT::Type{<:Annotation}, args::Annotation...)

Must return an AugmentedReturn type.

The primal must be the same type of the original return if needs_primal(config), otherwise nothing.

The shadow must be nothing if needs_shadow(config) is false. If width is 1, the shadow should be the same type of the original return. If the width is greater than 1, the shadow should be NTuple{original return, width}.

The tape can be any type (including Nothing) and is preserved for the reverse call.

EnzymeCore.EnzymeRules.forward — Function

forward(func::Annotation{typeof(f)}, RT::Type{<:Annotation}, args::Annotation...)

Calculate the forward derivative. The first argument func is the callable for which the rule applies to. Either wrapped in a Const), or a Duplicated if it is a closure. The second argument is the return type annotation, and all other arguments are the annotated function arguments.

EnzymeCore.EnzymeRules.inactive — Function

inactive(func::typeof(f), args...)

Mark a particular function as always being inactive in both its return result and the function call itself.

EnzymeCore.EnzymeRules.reverse — Function

reverse(::Config, func::Annotation{typeof(f)}, dret::Active, tape, args::Annotation...)
reverse(::Config, func::Annotation{typeof(f)}, ::Type{<:Annotation), tape, args::Annotation...)

Takes gradient of derivative, activity annotation, and tape. If there is an active return dret is passed as Active{T} with the derivative of the active return val. Otherwise dret is passed as Type{Duplicated{T}}, etc.