API

Types and constants

EnzymeCore.Active

EnzymeCore.Annotation

EnzymeCore.BatchDuplicated

EnzymeCore.BatchDuplicatedNoNeed

EnzymeCore.Const

EnzymeCore.Duplicated

EnzymeCore.DuplicatedNoNeed

EnzymeCore.ForwardMode

EnzymeCore.Mode

EnzymeCore.ReverseMode

Functions and macros

Enzyme.autodiff

Enzyme.autodiff

Enzyme.autodiff

Enzyme.autodiff_deferred

Enzyme.autodiff_deferred

Enzyme.fwddiff_deferred

Enzyme.fwddiff_deferred

Enzyme.gradient

Enzyme.gradient

Enzyme.gradient

Enzyme.gradient!

Enzyme.jacobian

Enzyme.jacobian

Documentation

Enzyme.autodiff — Method

autodiff(::ForwardMode, f, Activity, args...)

Auto-differentiate function f at arguments args using forward mode.

args may be numbers, arrays, structs of numbers, structs of arrays and so on. Enzyme will only differentiate in respect to arguments that are wrapped in a Duplicated or similar argument. Non-annotated arguments will automatically be treated as Const. Unlike reverse mode in autodiff, Active arguments are not allowed here, since all

Activity is the Activity of the return value, it may be:

Const if the return is not to be differentiated with respect to

Duplicated, if the return is being differentiated with respect to and both the original value and the derivative return are desired

DuplicatedNoNeed, if the return is being differentiated with respect to and only the derivative return is desired.

Example returning both original return and derivative:

a = 4.2
b = [2.2, 3.3]; ∂f_∂b = zero(b)
c = 55; d = 9

f(x) = x*x
res, ∂f_∂x = autodiff(Forward, f, Duplicated, Duplicated(3.14, 1.0))

# output

(9.8596, 6.28)

Example returning just the derivative:

a = 4.2
b = [2.2, 3.3]; ∂f_∂b = zero(b)
c = 55; d = 9

f(x) = x*x
∂f_∂x = autodiff(Forward, f, DuplicatedNoNeed, Duplicated(3.14, 1.0))

# output

(6.28,)

Enzyme.autodiff — Method

autodiff(::ReverseMode, f, Activity, args...)

Auto-differentiate function f at arguments args using reverse mode.

Limitations:

f may only return a Real (of a built-in/primitive type) or nothing, not an array, struct, BigFloat, etc. To handle vector-valued return types, use a mutating f! that returns nothing and stores it's return value in one of the arguments, which must be wrapped in a Duplicated.

args may be numbers, arrays, structs of numbers, structs of arrays and so on. Enzyme will only differentiate in respect to arguments that are wrapped in an Active (for immutable arguments like primitive types and structs thereof) or Duplicated (for mutable arguments like arrays, Refs and structs thereof). Non-annotated arguments will automatically be treated as Const.

Activity is the Activity of the return value, it may be Const or Active.

Example:

a = 4.2
b = [2.2, 3.3]; ∂f_∂b = zero(b)
c = 55; d = 9

f(a, b, c, d) = a * √(b[1]^2 + b[2]^2) + c^2 * d^2
∂f_∂a, _, _, ∂f_∂d = autodiff(Reverse, f, Active, Active(a), Duplicated(b, ∂f_∂b), c, Active(d))[1]

# output

(3.966106403010388, nothing, nothing, 54450.0)

here, autodiff returns a tuple $(\partial f/\partial a, \partial f/\partial d)$, while $\partial f/\partial b$ will be added to ∂f_∂b (but not returned). c will be treated as Const(c).

Note

Enzyme gradients with respect to integer values are zero. Active will automatically convert plain integers to floating point values, but cannot do so for integer values in tuples and structs.

Enzyme.autodiff — Method

autodiff(mode::Mode, f, args...)
autodiff(f, mode::Mode, args...)

Like autodiff but will try to guess the activity of the return value.

Enzyme.autodiff_deferred — Method

autodiff_deferred(f, Activity, args...)

Same as autodiff but uses deferred compilation to support usage in GPU code, as well as high-order differentiation.

Enzyme.autodiff_deferred — Method

autodiff_deferred(f, args...)

Like autodiff_deferred but will try to guess the activity of the return value.

Enzyme.fwddiff_deferred — Method

fwddiff_deferred(f, Activity, args...)

Same as autodiff(::ForwardMode, ...) but uses deferred compilation to support usage in GPU code, as well as high-order differentiation.

Enzyme.fwddiff_deferred — Method

fwddiff_deferred(f, args...)

Like fwddiff_deferred but will try to guess the activity of the return value.

Enzyme.gradient! — Method

gradient!(::ReverseMode, dx, f, x)

Compute the gradient of an array-input function f using reverse mode, storing the derivative result in an existing array dx.

Example:

f(x) = x[1]*x[2]

dx = [0.0, 0.0]
gradient!(Reverse, dx, f, [2.0, 3.0])

# output

2-element Vector{Float64}:
 3.0
 2.0

Enzyme.gradient — Method

gradient(::ForwardMode, f, x; shadow=onehot(x))

Compute the gradient of an array-input function f using forward mode. The optional keyword argument shadow is a vector of one-hot vectors of type x which are used to forward-propagate into the return. For performance reasons, this should be computed once, outside the call to gradient, rather than within this call.

Example:

f(x) = x[1]*x[2]

grad = gradient(Forward, f, [2.0, 3.0])

# output

(3.0, 2.0)

Enzyme.gradient — Method

gradient(::ReverseMode, f, x)

Compute the gradient of an array-input function f using reverse mode. This will allocate and return new array with the gradient result.

Example:

f(x) = x[1]*x[2]

grad = gradient(Reverse, f, [2.0, 3.0])

# output

2-element Vector{Float64}:
 3.0
 2.0

Enzyme.gradient — Method

gradient(::ForwardMode, f, x, ::Val{chunk}; shadow=onehot(x))

Compute the gradient of an array-input function f using vector forward mode. Like gradient, except it uses a chunk size of chunk to compute chunk derivatives in a single call.

Example:

f(x) = x[1]*x[2]

grad = gradient(Forward, f, [2.0, 3.0], Val(2))

# output

(3.0, 2.0)

Enzyme.jacobian — Method

jacobian(::ForwardMode, f, x; shadow=onehot(x))
jacobian(::ForwardMode, f, x, ::Val{chunk}; shadow=onehot(x))

Compute the jacobian of an array-input function f using (potentially vector) forward mode. This is a simple rename of the gradient function, and all relevant arguments apply here.

Example:

f(x) = [x[1]*x[2], x[2]]

grad = jacobian(Forward, f, [2.0, 3.0])

# output

2×2 Matrix{Float64}:
 3.0  2.0
 0.0  1.0

Enzyme.jacobian — Method

jacobian(::ReverseMode, f, x, ::Val{num_outs}, ::Val{chunk})

Compute the jacobian of an array-input function f using (potentially vector) reverse mode. The chunk argument denotes the chunk size to use and num_outs denotes the number of outputs f will return in an array.

Example:

f(x) = [x[1]*x[2], x[2]]

grad = jacobian(Reverse, f, [2.0, 3.0], Val(2))

# output

2×2 Matrix{Float64}:
 3.0  2.0
 0.0  1.0

EnzymeCore.Active — Type

Active(x)

Mark a function argument x of autodiff as active, Enzyme will auto-differentiate in respect Active arguments.

Note

Enzyme gradients with respect to integer values are zero. Active will automatically convert plain integers to floating point values, but cannot do so for integer values in tuples and structs.

EnzymeCore.Annotation — Type

abstract type Annotation{T}

Abstract type for autodiff function argument wrappers like Const, Active and Duplicated.

EnzymeCore.BatchDuplicated — Type

BatchDuplicated(x, ∂f_∂xs)

Like Duplicated, except contains several shadows to compute derivatives for all at once. Argument ∂f_∂xs should be a tuple of the several values of type x.

EnzymeCore.BatchDuplicatedNoNeed — Type

BatchDuplicatedNoNeed(x, ∂f_∂xs)

Like DuplicatedNoNeed, except contains several shadows to compute derivatives for all at once. Argument ∂f_∂xs should be a tuple of the several values of type x.

EnzymeCore.Const — Type

Const(x)

Mark a function argument x of autodiff as constant, Enzyme will not auto-differentiate in respect Const arguments.

EnzymeCore.Duplicated — Type

Duplicated(x, ∂f_∂x)

Mark a function argument x of autodiff as duplicated, Enzyme will auto-differentiate in respect to such arguments, with dx acting as an accumulator for gradients (so $\partial f / \partial x$ will be added to) ∂f_∂x.

EnzymeCore.DuplicatedNoNeed — Type

DuplicatedNoNeed(x, ∂f_∂x)

Like Duplicated, except also specifies that Enzyme may avoid computing the original result and only compute the derivative values.

EnzymeCore.ForwardMode — Type

struct Forward <: Mode

Forward mode differentiation

EnzymeCore.Mode — Type

abstract type Mode

Abstract type for what differentiation mode will be used.

EnzymeCore.ReverseMode — Type

struct Reverse <: Mode

Reverse mode differentiation