Internal API

Note

This is the documentation of Enzymes's internal API. The internal API is not subject to semantic versioning and may change at any time and without deprecation.

Enzyme.Compiler.fspec — Method

Create the methodinstance pair, and lookup the primal return type.